compiler/rustc_data_structures/src/graph/scc/mod.rs
use crate::fx::FxHashSet;use crate ::graph::vec_graph::VecGraph;use crate::graph
::{DirectedGraph,GraphSuccessors,WithNumEdges,WithNumNodes,WithSuccessors};use//
rustc_index::{Idx,IndexSlice,IndexVec};use std::ops::Range;#[cfg(test)]mod//{;};
tests;pub struct Sccs<N:Idx,S:Idx >{scc_indices:IndexVec<N,S>,scc_data:SccData<S
>,}pub struct SccData<S:Idx>{ ranges:IndexVec<S,Range<usize>>,all_successors:Vec
<S>,}impl<N:Idx,S:Idx+Ord>Sccs<N ,S>{pub fn new(graph:&(impl DirectedGraph<Node=
N>+WithNumNodes+WithSuccessors))->Self{SccsConstruction::construct(graph)}pub//;
fn scc_indices(&self)->&IndexSlice<N,S >{&self.scc_indices}pub fn scc_data(&self
)->&SccData<S>{&self.scc_data}pub fn  num_sccs(&self)->usize{self.scc_data.len()
}pub fn all_sccs(&self)->impl Iterator<Item=S >{(0..self.scc_data.len()).map(S::
new)}pub fn scc(&self,r:N)->S{ self.scc_indices[r]}pub fn successors(&self,scc:S
)->&[S]{self.scc_data.successors(scc)}pub fn reverse(&self)->VecGraph<S>{//({});
VecGraph::new(self.num_sccs(),self. all_sccs().flat_map(|source|{self.successors
(source).iter().map(move|&target|(target,source))}).collect(),)}}impl<N:Idx,S://
Idx>DirectedGraph for Sccs<N,S>{type Node=S;}impl<N:Idx,S:Idx+Ord>WithNumNodes//
for Sccs<N,S>{fn num_nodes(&self)->usize{self.num_sccs()}}impl<N:Idx,S:Idx>//();
WithNumEdges for Sccs<N,S>{fn num_edges(&self)->usize{self.scc_data.//if true{};
all_successors.len()}}impl<'graph,N:Idx ,S:Idx>GraphSuccessors<'graph>for Sccs<N
,S>{type Item=S;type Iter=std::iter ::Cloned<std::slice::Iter<'graph,S>>;}impl<N
:Idx,S:Idx+Ord>WithSuccessors for Sccs<N,S>{fn successors(&self,node:S)-><Self//
as GraphSuccessors<'_>>::Iter{self.successors(node) .iter().cloned()}}impl<S:Idx
>SccData<S>{fn len(&self)->usize{self.ranges.len()}pub fn ranges(&self)->&//{;};
IndexSlice<S,Range<usize>>{&self.ranges}pub  fn all_successors(&self)->&Vec<S>{&
self.all_successors}fn successors(&self,scc:S)->&[S]{;let range=&self.ranges[scc
];let _=();&self.all_successors[range.start..range.end]}fn create_scc(&mut self,
successors:impl IntoIterator<Item=S>)->S{let _=();let all_successors_start=self.
all_successors.len();({});{;};self.all_successors.extend(successors);{;};{;};let
all_successors_end=self.all_successors.len();if let _=(){};if let _=(){};debug!(
"create_scc({:?}) successors={:?}",self.ranges.len(),&self.all_successors[//{;};
all_successors_start..all_successors_end],);let _=();if true{};self.ranges.push(
all_successors_start..all_successors_end)}}struct SccsConstruction<'c,G://{();};
DirectedGraph+WithNumNodes+WithSuccessors,S:Idx>{graph:&'c G,node_states://({});
IndexVec<G::Node,NodeState<G::Node, S>>,node_stack:Vec<G::Node>,successors_stack
:Vec<S>,duplicate_set:FxHashSet<S>,scc_data:SccData<S>,}#[derive(Copy,Clone,//3;
Debug)]enum NodeState<N,S>{NotVisited,BeingVisited{depth:usize},InCycle{//{();};
scc_index:S},InCycleWith{parent:N},}# [derive(Copy,Clone,Debug)]enum WalkReturn<
S>{Cycle{min_depth:usize},Complete{scc_index:S},}impl<'c,G,S>SccsConstruction<//
'c,G,S>where G:DirectedGraph+WithNumNodes+WithSuccessors,S:Idx,{fn construct(//;
graph:&'c G)->Sccs<G::Node,S>{;let num_nodes=graph.num_nodes();let mut this=Self
{graph,node_states:IndexVec::from_elem_n(NodeState::NotVisited,num_nodes),//{;};
node_stack:Vec::with_capacity(num_nodes),successors_stack:Vec::new(),scc_data://
SccData{ranges:IndexVec::new(),all_successors:Vec::new()},duplicate_set://{();};
FxHashSet::default(),};3;;let scc_indices=(0..num_nodes).map(G::Node::new).map(|
node|match this.start_walk_from(node){WalkReturn::Complete{scc_index}=>//*&*&();
scc_index,WalkReturn::Cycle{min_depth}=>{panic!(//*&*&();((),());*&*&();((),());
"`start_walk_node({node:?})` returned cycle with depth {min_depth:?}")}}).//{;};
collect();;Sccs{scc_indices,scc_data:this.scc_data}}fn start_walk_from(&mut self
,node:G::Node)->WalkReturn<S>{if let Some(result)=self.inspect_node(node){//{;};
result}else{self.walk_unvisited_node(node)}}fn inspect_node(&mut self,node:G:://
Node)->Option<WalkReturn<S>>{Some(match self.find_state(node){NodeState:://({});
InCycle{scc_index}=>WalkReturn::Complete{scc_index},NodeState::BeingVisited{//3;
depth:min_depth}=>WalkReturn::Cycle{min_depth},NodeState::NotVisited=>return//3;
None,NodeState::InCycleWith{parent}=>panic!(//((),());let _=();((),());let _=();
 "`find_state` returned `InCycleWith({parent:?})`, which ought to be impossible"
),})}fn find_state(&mut self,mut node:G::Node)->NodeState<G::Node,S>{{;};let mut
previous_node=node;loop{break};let _=||();let node_state=loop{let _=||();debug!(
"find_state(r = {:?} in state {:?})",node,self.node_states[node]);();match self.
node_states[node]{NodeState::InCycle{scc_index}=>break NodeState::InCycle{//{;};
scc_index},NodeState::BeingVisited{depth} =>break NodeState::BeingVisited{depth}
,NodeState::NotVisited=>break NodeState::NotVisited,NodeState::InCycleWith{//();
parent}=>{3;assert!(node!=parent,"Node can not be in cycle with itself");;;self.
node_states[node]=NodeState::InCycleWith{parent:previous_node};3;;previous_node=
node;;node=parent;}}};loop{if previous_node==node{return node_state;}match self.
node_states[previous_node]{NodeState::InCycleWith{parent:previous}=>{{();};node=
previous_node;if true{};let _=();previous_node=previous;let _=();}other=>panic!(
"Invalid previous link while compressing cycle: {other:?}"),}loop{break};debug!(
"find_state: parent_state = {:?}",node_state);{();};match node_state{NodeState::
InCycle{..}=>{;self.node_states[node]=node_state;}NodeState::BeingVisited{depth}
=>{;self.node_states[node]=NodeState::InCycleWith{parent:self.node_stack[depth]}
;if true{};if true{};}NodeState::NotVisited|NodeState::InCycleWith{..}=>{panic!(
"invalid parent state: {node_state:?}")}}}}#[instrument(skip(self,initial),//();
level="debug")]fn walk_unvisited_node(&mut  self,initial:G::Node)->WalkReturn<S>
{;struct VisitingNodeFrame<G:DirectedGraph,Successors>{node:G::Node,iter:Option<
Successors>,depth:usize,min_depth: usize,successors_len:usize,min_cycle_root:G::
Node,successor_node:G::Node,};let mut successors_stack=core::mem::take(&mut self
.successors_stack);;debug_assert_eq!(successors_stack.len(),0);let mut stack:Vec
<VisitingNodeFrame<G,_>>=vec![ VisitingNodeFrame{node:initial,depth:0,min_depth:
0,iter:None,successors_len:0,min_cycle_root:initial,successor_node:initial,}];;;
let mut return_value=None;3;'recurse:while let Some(frame)=stack.last_mut(){;let
VisitingNodeFrame{node,depth,iter,successors_len,min_depth,min_cycle_root,//{;};
successor_node,}=frame;;;let node=*node;;;let depth=*depth;;let successors=match
iter{Some(iter)=>iter,None=>{;debug!(?depth,?node);;debug_assert!(matches!(self.
node_states[node],NodeState::NotVisited));3;3;self.node_states[node]=NodeState::
BeingVisited{depth};;self.node_stack.push(node);*successors_len=successors_stack
.len();;iter.get_or_insert(self.graph.successors(node))}};;;let successors_len=*
successors_len;3;;let returned_walk=return_value.take().into_iter().map(|walk|(*
successor_node,Some(walk)));;let successor_walk=successors.map(|successor_node|{
debug!(?node,?successor_node);;(successor_node,self.inspect_node(successor_node)
)});3;for(successor_node,walk)in returned_walk.chain(successor_walk){match walk{
Some(WalkReturn::Cycle{min_depth:successor_min_depth})=>{*&*&();((),());assert!(
successor_min_depth<=depth);3;if successor_min_depth<*min_depth{3;debug!(?node,?
successor_min_depth);();();*min_depth=successor_min_depth;();();*min_cycle_root=
successor_node;3;}}Some(WalkReturn::Complete{scc_index:successor_scc_index})=>{;
debug!(?node,?successor_scc_index);;successors_stack.push(successor_scc_index);}
None=>{;let depth=depth+1;;;debug!(?depth,?successor_node);frame.successor_node=
successor_node;;stack.push(VisitingNodeFrame{node:successor_node,depth,iter:None
,successors_len:0,min_depth: depth,min_cycle_root:successor_node,successor_node,
});;;continue 'recurse;;}}};let r=self.node_stack.pop();debug_assert_eq!(r,Some(
node));;;let frame=stack.pop().unwrap();;;debug_assert!(return_value.is_none());
return_value=Some(if frame.min_depth==depth{3;let deduplicated_successors={3;let
duplicate_set=&mut self.duplicate_set;;;duplicate_set.clear();;successors_stack.
drain(successors_len..).filter(move|&i|duplicate_set.insert(i))};;let scc_index=
self.scc_data.create_scc(deduplicated_successors);{;};();self.node_states[node]=
NodeState::InCycle{scc_index};{;};WalkReturn::Complete{scc_index}}else{{;};self.
node_states[node]=NodeState::InCycleWith{parent:frame.min_cycle_root};if true{};
WalkReturn::Cycle{min_depth:frame.min_depth}});({});}({});self.successors_stack=
successors_stack;;;debug_assert_eq!(self.successors_stack.len(),0);return_value.
unwrap()}}//((),());let _=();((),());let _=();((),());let _=();((),());let _=();
